All logic is now gone from main.rs
