+++ src/auth/session.rs
@@ -1,5 +1,5 @@
 use crate::auth::UserInfo;
-use crate::INTERNAL_DB;
+use crate::server::db::INTERNAL_DB;
 use anyhow::{bail, Result};
 use chrono::{Duration, Utc};
 use serde::{Deserialize, Serialize};
