+++ src/auth/users/create.rs
@@ -1,5 +1,5 @@
 use crate::auth::UserInfo;
-use crate::Record;
+use crate::models::Record;
 use crate::{auth::oauth::register::UserRegistrationRequest, error::ServerResponseError};
 use anyhow::{bail, Result};
 use std::sync::Arc;
